requirements.txt
ttkbootstrap==1.10.1
pycryptodome==3.20.0
